requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
# APScheduler==3.8.1
args==0.1.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
boto3==1.21.3
botocore==1.24.3
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
clint==0.5.1
coverage==4.0.3
coveralls==1.1
cryptography==36.0.1
curdling==0.4.0
defusedxml==0.7.1
distlib==0.1.2
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.47.0
django-appconf==1.0.5
django-ckeditor==6.1.0
django-cors-headers==3.10.0
# django-countries==7.2.1
# django-crontab==0.7.1
django-cuser==2017.3.16
django-debug-toolbar==3.2.4
# django-excel==0.0.10
# django-hosts==5.0
django-imagekit==4.1.0
django-ipware==4.0.2
django-js-asset==1.2.2
# django-mathfilters==1.0.0
django-qr-code==3.0.0
# django-static-fontawesome==5.14.0.0
django-storages==1.12.3
# django-tinymce==3.4.0
# django-widget-tweaks==1.4.9
# django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.12
drf-nested-routers==0.93.4
frozenlist==1.2.0
funcsigs==0.4
geoip2==4.5.0
gunicorn==20.1.0
httpretty==0.8.14
idna==2.7
jmespath==0.10.0
lml==0.1.0
m2r==0.2.1
maxminddb==2.2.0
mistune==0.8.4
mock==1.3.0
multidict==5.2.0
oauthlib==3.1.1
# paystackapi==2.0.0
pbr==1.8.1
pilkit==2.0
Pillow==8.3.2
pkginfo==1.4.2
psycopg2==2.9.3
pycparser==2.21
# pyexcel==0.6.7
# pyexcel-io==0.6.5
# pyexcel-webio==0.1.4
# PyJWT==2.3.0
python-dateutil==2.8.2
# python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
s3transfer==0.5.1
segno==1.4.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.4.2
sure==1.2.24
texttable==1.6.4
tqdm==4.25.0
twine==1.11.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
whitenoise==5.3.0
# wkhtmltopdf==0.2
# yarl==1.7.2
